rearranged Chest classes, make DeathChest class
