+++ src/com/mojang/ld22/entity/DungeonChest.java
@@ -12,18 +12,17 @@ import com.mojang.ld22.screen.ContainerMenu;
 import com.mojang.ld22.screen.ModeMenu;
 import java.util.Random;
 
-public class DungeonChest extends Furniture {
+public class DungeonChest extends Chest {
 
-	public Inventory inventory;
 	public Random random = new Random();
-	public boolean islocked;
+	public boolean isLocked;
 	private static int openCol = Color.get(-1, 2, 115, 225);
 	
 	public DungeonChest() {
 		super("Dungeon Chest");
 		populateInv();
 		
-		islocked = true;
+		isLocked = true;
 		sprite = 1;
 		col = Color.get(-1, 222, 333, 555);
 		col0 = Color.get(-1, 111, 222, 444);
@@ -33,48 +32,54 @@ public class DungeonChest extends Furniture {
 	}
 	
 	public boolean use(Player player, int attackDir) {
-		if(!islocked) {
-			player.game.setMenu(new ContainerMenu(player, name, inventory));
-			return true;
-		} else {
+		/*if(!isLocked) {
+			//player.game.setMenu(new ContainerMenu(player, name, inventory));
+			return super.use(player, attackDir);
+		} else*/if (isLocked) {
 			boolean activeKey = player.activeItem != null && player.activeItem.getName().equals("Key");
 			boolean invKey = player.inventory.hasResources(Resource.key, 1);
-			if(activeKey || invKey) {
-				if (!ModeMenu.creative) {
-					if (activeKey) {
+			if(activeKey || invKey) { // if the player has a key...
+				if (!ModeMenu.creative) { // remove the key unless on creative mode.
+					if (activeKey) { // remove activeItem
 						ResourceItem key = (ResourceItem)player.activeItem;
 						key.count--;
-					} else {
+					} else { // remove from inv
 						player.inventory.removeResource(Resource.key, 1);
 					}
 				}
 				
-				islocked = false;
-				player.game.setMenu(new ContainerMenu(player, name, inventory));
-				col = col0 = col1 = col2 = col3 = openCol;
+				isLocked = false;
+				//player.game.setMenu(new ContainerMenu(player, name, inventory));
+				col = col0 = col1 = col2 = col3 = openCol; // set to the unlocked color
 				
 				level.add(new SmashParticle(x * 16 + 8, y * 16 + 8));
 				level.add(new TextParticle("-1 key", x, y, Color.get(-1, 500, 500, 500)));
 				level.chestcount--;
-				if(level.chestcount == 0) {
-					for(int i = 0; i < 5; i++) {
+				if(level.chestcount == 0) { // if this was the last chest...
+					for(int i = 0; i < 5; i++) { // add 5 golden apples to the level
 						level.add(new ItemEntity(new ResourceItem(Resource.goldapple), x, y));
 					}
 					
-					Game.notifications.add("You hear a noise from the surface!");
+					Game.notifications.add("You hear a noise from the surface!"); // notify the player of the developments
+					// add a level 2 airwizard to the middle surface level.
 					AirWizard wizard = new AirWizard(true);
 					wizard.x = Game.levels[3].w / 2;
 					wizard.y = Game.levels[3].h / 2;
 					Game.levels[3].add(wizard);
 				}
+				
+				return super.use(player, attackDir); // the player unlocked the chest.
 			}
 			
-			return true;
+			return false; // the chest is locked, and the player has no key.
 		}
+		else return super.use(player, attackDir); // the chest was already unlocked.
 	}
 	
+	/** Populate the inventory of the DungeonChest, psudo-randomly. */
 	private void populateInv() {
-		inventory = new Inventory(this);
+		//inventory = new Inventory();
+		inventory.clearInv();
 		tryAdd(5, Resource.steak, 6);
 		tryAdd(5, Resource.cookedpork, 6);
 		tryAdd(4, Resource.wood, 20);
@@ -104,13 +109,14 @@ public class DungeonChest extends Furniture {
 		tryAdd(4, ToolType.claymore, 1);
 		tryAdd(6, ToolType.claymore, 2);
 		
-		if(inventory.invSize() < 1) {
+		if(inventory.invSize() < 1) { // add this if none of the above was added.
 			inventory.add(new ResourceItem(Resource.steak, 6));
 			inventory.add(new ResourceItem(Resource.timepotion, 1));
-			inventory.add(new ToolItem(ToolType.hatchet, 4));
+			inventory.add(new ToolItem(ToolType.axe, 4));
 		}
 	}
 	
+	/** The functions that (possibly) add Resources and Tools to the chest, respectively. */
 	private void tryAdd(int chance, Resource item, int num) {
 		if(random.nextInt(chance) == 1)
 			inventory.add(new ResourceItem(item, num));
@@ -119,19 +125,21 @@ public class DungeonChest extends Furniture {
 		if(random.nextInt(chance) == 1)
 			inventory.add(new ToolItem(type, lvl));
 	}
-
+	
+	/** what happens if the player tries to push a Dungeon Chest. */
 	protected void touchedBy(Entity entity) {
-		if(entity instanceof Player && pushTime == 0 && !islocked) {
+		/*if(entity instanceof Player && pushTime == 0 && !isLocked) {
 			pushDir = ((Player)entity).dir;
 			pushTime = 10;
-		}
-
+		}*/
+		if(!isLocked) super.touchedBy(entity);
 	}
-
+	
+	/** what happens if the player tries to grab a Dungeon Chest. */
 	public void take(Player player) {
-		if(!islocked) {
+		/*if(!isLocked) {
 			shouldTake = player;
-		}
-
+		}*/
+		if(!isLocked) super.take(player);
 	}
 }
